add conversion support for 'NettyProvider'
